Disabled Hydro
Edited the init.lua file to remove the requirement of water for crops to
grow if set in the settings. Did this by adding a local variable called
hydro, and if set to "on" in the config file code was executed that
affected the water, and if set to "off did not". Mainly done for user
customizability.